requirements.txt
Flask==0.12.2

gunicorn==19.7.1

itsdangerous==0.24

Jinja2==2.10
MarkupSafe==1.0

requests==2.18.1

six==1.10.0

urllib3==1.21.1

Werkzeug==0.14.1
click==6.7
pip==9.0.3
setuptools==39.0.1
line-bot-sdk==1.1.0
pycurl==7.43.0.1
numpy==1.14.2 
opencv-python==3.4.0.12
opencv-contrib-python==3.4.0.12
matplotlib==2.2.2
Pillow==5.1.0
colormath==3.0.0